StratForge/services/worker-python/requirements.txt
redis==5.0.1
pandas==2.2.0
python-dotenv==1.0.1
